+++ packages/react-component-library/src/components/NumberInputE/useEarlyValidation.ts
@@ -1,15 +1,44 @@
 import React, { useCallback } from 'react'
 
-import { areCharactersValid } from './validation'
+import { isValueValid } from './validation'
 
-export function useEarlyValidation(): {
-  handleKeyDown: (event: React.KeyboardEvent<HTMLInputElement>) => void
+/**
+ * Given some text that is being entered in an input, work out what the
+ * new value will be based on the cursor position ond selection.
+ */
+function getNewInputValue(input: HTMLInputElement, newText: string): string {
+  const textBeforeSelection = input.value.substring(0, input.selectionStart)
+  const textAfterSelection = input.value.substring(input.selectionEnd)
+  return `${textBeforeSelection}${newText}${textAfterSelection}`
+}
+
+export function useEarlyValidation(isNegativeAllowed: boolean): {
   handleBeforeInput: (
     event:
       | React.FormEvent<HTMLInputElement>
       | React.CompositionEvent<HTMLInputElement>
   ) => void
+  handlePaste: (event: React.ClipboardEvent<HTMLInputElement>) => void
 } {
+  /**
+   * Block invalid pastes before the DOM is updated (to
+   * avoid the cursor jumping).
+   *
+   * Only needed for Firefox and IE.
+   */
+  const handlePaste = useCallback(
+    (event: React.ClipboardEvent<HTMLInputElement>) => {
+      const pastedText = event.clipboardData.getData('text')
+      const newValue = getNewInputValue(event.currentTarget, pastedText)
+
+      if (!isValueValid(newValue, isNegativeAllowed)) {
+        event.preventDefault()
+        event.stopPropagation()
+      }
+    },
+    [isNegativeAllowed]
+  )
+
   /**
    * Block invalid input before the DOM is updated where possible (to
    * avoid the cursor jumping).
@@ -33,32 +62,18 @@ export function useEarlyValidation(): {
       // replacement (if existing text was selected).
       const { data } = event
 
-      if (!areCharactersValid(data)) {
+      const newValue = getNewInputValue(event.currentTarget, data)
+
+      if (!isValueValid(newValue, isNegativeAllowed)) {
         event.preventDefault()
         event.stopPropagation()
       }
     },
-    []
-  )
-
-  /**
-   * Block a second decimal point from being typed (before the DOM is
-   * updated, to avoid the cursor jumping).
-   */
-  const handleKeyDown = useCallback(
-    (event: React.KeyboardEvent<HTMLInputElement>) => {
-      const currentValue = event.currentTarget.value
-
-      if (event.key === '.' && currentValue.includes('.')) {
-        event.preventDefault()
-        event.stopPropagation()
-      }
-    },
-    []
+    [isNegativeAllowed]
   )
 
   return {
     handleBeforeInput,
-    handleKeyDown,
+    handlePaste,
   }
 }