+++ packages/icon-library/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.79.1](https://github.com/defencedigital/mod-uk-design-system/compare/2.79.0...2.79.1) (2022-01-28)
+
+**Note:** Version bump only for package @defencedigital/icon-library
+
+
+
+
+
+# [2.79.0](https://github.com/defencedigital/mod-uk-design-system/compare/2.78.2...2.79.0) (2022-01-27)
+
+**Note:** Version bump only for package @defencedigital/icon-library
+
+
+
+
+
 ## [2.78.2](https://github.com/defencedigital/mod-uk-design-system/compare/2.78.1...2.78.2) (2022-01-25)
 
 **Note:** Version bump only for package @defencedigital/icon-library